turbo-engine/src/main/java/com/xiaoju/uemc/turbo/engine/dao/NodeInstanceLogDAO.java
package com.xiaoju.uemc.turbo.engine.dao;

import com.xiaoju.uemc.turbo.engine.dao.mapper.NodeInstanceLogMapper;
import com.xiaoju.uemc.turbo.engine.entity.NodeInstanceLogPO;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by Stefanie on 2019/12/5.
 */
@Service
public class NodeInstanceLogDAO extends BaseDAO<NodeInstanceLogMapper, NodeInstanceLogPO> {

    /**
     * insert nodeInstanceLogPO
     * if error, this will not throw exption but return -1
     *
     * @param nodeInstanceLogPO
     * @return
     */
    public int insert(NodeInstanceLogPO nodeInstanceLogPO) {
        try {
            return baseMapper.insert(nodeInstanceLogPO);
        } catch (Exception e) {
            LOGGER.error("insert exception.||nodeInstanceLogPO={}", nodeInstanceLogPO, e);
        }
        return -1;
    }

    /**
     * batch insert nodeInstanceLogList
     *
     * @param nodeInstanceLogList
     * @return
     */
    public boolean insertList(List<NodeInstanceLogPO> nodeInstanceLogList) {
        return baseMapper.batchInsert(nodeInstanceLogList.get(0).getFlowInstanceId(), nodeInstanceLogList);
    }
}
